requirements.txt
fastapi
uvicorn
pydantic
python-multipart
jinja2
python-dotenv
groq
langchain
langchain-community
chromadb
sentence-transformers
arxiv
requests
pypdf
pdfplumber
PyMuPDF
networkx
matplotlib
pandas
numpy
beautifulsoup4
plotly
wordcloud
torch
aiofiles
watchdog
seaborn
PyJWT
flask